Updates SERA craton GMPE to new NGAEast/USGS CEUS models
